+++ packages/atlas/src/components/TopSellingChannelsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './TopSellingChannelsTable'
